requirements.txt
alembic==1.13.3
alembic-postgresql-enum==1.8.0
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.30.0
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
email_validator==2.2.0
fastapi==0.119.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
graypy==2.1.0
resend==2.19.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.3.5
pandas==2.3.3
pillow==11.0.0
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.8.0
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.19
pytz==2025.2
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
sentry-sdk==2.42.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.48.0
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.32.1
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
xlsxwriter==3.2.9
openpyxl==3.1.5
